+++ lib/presentation/app.dart
@@ -12,8 +12,8 @@ import '../utils/extensions.dart';
 import '../utils/localizations/strings.g.dart';
 import '../utils/logger.dart';
 import '../utils/routing/router.dart';
+import 'components/states/url_launcher.dart';
 import 'components/theme.dart';
-import 'components/url_launcher.dart';
 
 /// GitHubSearch アプリ
 class GitHubSearchApp extends StatelessWidget {
